+++ CHANGELOG.md
@@ -45,6 +45,7 @@ All notable changes to this project will be documented in this file.
 - [Core] Get KSP2 version from game assembly (#4034 by: HebaruSan)
 - [Multiple] Build nuget package, support netstandard2.0 build (#4039 by: HebaruSan)
 - [Core] Use fully sanitized archive.org bucket names (#4043 by: HebaruSan)
+- [Netkan] Omit duplicate inflation warnings in queue (#4071 by: HebaruSan)
 
 ## v1.34.4 (Niven)